% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars-adminlte.R
\name{adminlte_vars}
\alias{adminlte_vars}
\title{AdminLTE 2 custom variables}
\usage{
adminlte_vars(...)
}
\arguments{
\item{...}{Variables to use, under the form \code{body_bg = "#FFF"} or \code{"body-bg" = "#FFF"}.}
}
\value{
a \code{list} that can be used in \code{\link{create_theme}}.
}
\description{
Use any AdminLTE or Bootstrap variables to customize a \{shinydashboard\} theme.
}
\note{
For a full list of available variables, use \code{\link{search_vars_adminlte2}}.
}
\examples{
adminlte_vars(body_bg = "#FFF")

adminlte_vars("body-bg" = "#FFF")
}
