% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars-bs4dash.R
\name{bs4dash_vars}
\alias{bs4dash_vars}
\title{bs4dash custom variables}
\usage{
bs4dash_vars(...)
}
\arguments{
\item{...}{Variables to use, under the form \code{body_bg = "#FFF"} or \code{"body-bg" = "#FFF"}.}
}
\value{
a \code{list} that can be used in \code{\link{create_theme}}.
}
\description{
Use any AdminLTE or Bootstrap variables to customize a \{bs4Dash\} theme.
}
\note{
For a full list of available variables, use \code{\link{search_bs4dash_vars}}.
}
\examples{
bs4dash_vars(body_bg = "#FFF")

bs4dash_vars("body-bg" = "#FFF")
}
