% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars-bootstrap.R
\name{bs_vars}
\alias{bs_vars}
\title{Bootstrap custom variables}
\usage{
bs_vars(...)
}
\arguments{
\item{...}{Variables to use, under the form \code{body_bg = "#FFF"} or \code{"body-bg" = "#FFF"}.}
}
\value{
a \code{list} that can be used in \code{\link{create_theme}}.
}
\description{
Use any Bootstrap variables to customize a \{shiny\} theme.
}
\note{
For a full list of available variables, use \code{\link{search_vars_bs}}.
}
\examples{
bs_vars(body_bg = "#FFF")

bs_vars("body-bg" = "#FFF")
}
