% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars-scss.R
\name{bs_vars_color}
\alias{bs_vars_color}
\title{Bootstrap CSS colors variables}
\usage{
bs_vars_color(brand_primary = NULL, brand_success = NULL,
  brand_info = NULL, brand_warning = NULL, brand_danger = NULL,
  body_bg = NULL, text_color = NULL, gray_base = NULL,
  gray_darker = NULL, gray_dark = NULL, gray = NULL,
  gray_light = NULL, gray_lighter = NULL)
}
\arguments{
\item{brand_primary}{Primary color (default in Bootstrap is blue).}

\item{brand_success}{Success color (default in Bootstrap is green).}

\item{brand_info}{Info color (default in Bootstrap is light blue).}

\item{brand_warning}{Warning color (default in Bootstrap is orange).}

\item{brand_danger}{Danger color (default in Bootstrap is red).}

\item{body_bg}{Background color (default in Bootstrap is white).}

\item{text_color}{Text color (default in Bootstrap is black).}

\item{gray_base}{Base gray color.}

\item{gray_darker}{Darker gray color.}

\item{gray_dark}{Dark gray color.}

\item{gray}{Gray color.}

\item{gray_light}{Light gray color.}

\item{gray_lighter}{Lighter gray color.}
}
\value{
a \code{list} that can be used in \code{\link{create_theme}}.
}
\description{
Those variables can be used to customize Bootstrap and Bootswatch themes.
}
\note{
See default parameters for Bootstrap: \url{https://getbootstrap.com/docs/3.4/customize/}.
}
\examples{

# New colors (for buttons for example)
bs_vars_color(
  brand_primary = "#75b8d1",
  brand_success = "#c9d175",
  brand_info = "#758bd1",
  brand_warning = "#d1ab75",
  brand_danger = "#d175b8"
)
}
