% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fresh-package.R
\docType{package}
\name{fresh-package}
\alias{fresh}
\alias{fresh-package}
\title{fresh: FWA-Referenced Spatial Hydrology}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Stream network-aware spatial operations via direct SQL against fwapg and bcfishpass. Snap points to streams, delineate watersheds, query fish observations, and fetch network-referenced lakes, wetlands, and streams from the BC Freshwater Atlas. Requires PostgreSQL with fwapg, bcfishpass, and bcfishobs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NewGraphEnvironment/fresh}
  \item \url{https://newgraphenvironment.github.io/fresh/}
  \item Report bugs at \url{https://github.com/NewGraphEnvironment/fresh/issues}
}

}
\author{
\strong{Maintainer}: Allan Irvine \email{al@newgraphenvironment.com} (\href{https://orcid.org/0000-0002-3495-2128}{ORCID})

Other contributors:
\itemize{
  \item New Graph Environment [copyright holder]
}

}
\keyword{internal}
