% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fresh-package.R
\docType{package}
\name{fresh}
\alias{fresh-package}
\alias{fresh}
\title{Fresh 'Shiny' Themes}
\description{
Customize 'Bootstrap' and 'Bootswatch' themes, like colors, fonts, grid layout, to use in 'Shiny' applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dreamRs/fresh}
  \item Report bugs at \url{https://github.com/dreamRs/fresh/issues}
}

}
\author{
Victor Perrier & Fanny Meyer (\href{https://twitter.com/dreamRs_fr}{@dreamRs_fr})
}
