% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_db_query.R
\name{frs_db_query}
\alias{frs_db_query}
\title{Query FWA PostgreSQL Database}
\usage{
frs_db_query(conn, query)
}
\arguments{
\item{conn}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object (from \code{\link[=frs_db_conn]{frs_db_conn()}}).}

\item{query}{Character. SQL query string.}
}
\value{
An \code{sf} data frame (if the query returns geometry) or a plain
data frame.
}
\description{
Executes a SQL query on an open connection. Uses \code{\link[sf:st_read]{sf::st_read()}} so
spatial columns are returned as sf geometry.
}
\examples{
\dontrun{
conn <- frs_db_conn()
frs_db_query(conn, "SELECT * FROM whse_basemapping.fwa_lakes_poly LIMIT 5")
DBI::dbDisconnect(conn)
}
}
\seealso{
Other database: 
\code{\link{frs_db_conn}()}
}
\concept{database}
