% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_fish_obs.R
\name{frs_fish_obs}
\alias{frs_fish_obs}
\title{Fetch Fish Observations}
\usage{
frs_fish_obs(
  species_code = NULL,
  watershed_group_code = NULL,
  blue_line_key = NULL,
  limit = NULL,
  ...
)
}
\arguments{
\item{species_code}{Character. Species code (e.g. \code{"CH"} for chinook,
\code{"ST"} for steelhead). Default \code{NULL} (all species).}

\item{watershed_group_code}{Character. Watershed group code. Default \code{NULL}.}

\item{blue_line_key}{Integer. Blue line key. Default \code{NULL}.}

\item{limit}{Integer. Maximum rows to return. Default \code{NULL}.}

\item{...}{Additional arguments passed to \code{\link[=frs_db_conn]{frs_db_conn()}}.}
}
\value{
An \code{sf} data frame of fish observation events.
}
\description{
Query fish observation events from \code{bcfishobs.fiss_fish_obsrvtn_events_vw}.
Filter by species code, watershed group, and/or blue line key.
}
\examples{
\dontrun{
# Chinook observations in the Bulkley
obs <- frs_fish_obs(species_code = "CH", watershed_group_code = "BULK")
}
}
\seealso{
Other fish: 
\code{\link{frs_fish_habitat}()}
}
\concept{fish}
