% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_vars.R
\name{search_bs4dash_vars}
\alias{search_bs4dash_vars}
\title{Search bs4Dash variables}
\usage{
search_bs4dash_vars(pattern = NULL, source = c("adminlte", "bootstrap"))
}
\arguments{
\item{pattern}{A pattern to filter the results.}

\item{source}{Search variables in AdminLTE or Bootstrap or both.}
}
\value{
a \code{data.frame}
}
\description{
Search bs4Dash variables
}
\examples{
# Retrieve all variables
all_vars <- search_bs4dash_vars()
head(all_vars, 20)

# Search for a pattern
head(search_bs4dash_vars("navbar"))
}
