% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-vars.R
\name{search_vars_adminlte2}
\alias{search_vars_adminlte2}
\title{Search AdminLTE 2 shinydashboard variables}
\usage{
search_vars_adminlte2(pattern = NULL)
}
\arguments{
\item{pattern}{A pattern to filter the results.}
}
\value{
a \code{data.frame} with two columns:
\itemize{
\item \code{variable} : name of the variable.
\item \code{value} default value used.
}
}
\description{
Search AdminLTE 2 shinydashboard variables
}
\examples{
# All AdminLTE2 variables
search_vars_adminlte2()

# Only sidebar related variables
search_vars_adminlte2(pattern = "sidebar")
}
