% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-vars.R
\name{search_vars_bs}
\alias{search_vars_bs}
\title{Search Bootstrap variables}
\usage{
search_vars_bs(
  pattern = NULL,
  theme = c("default", "cerulean", "cosmo", "cyborg", "darkly", "flatly", "journal",
    "lumen", "paper", "readable", "sandstone", "simplex", "slate", "spacelab",
    "superhero", "united", "yeti")
)
}
\arguments{
\item{pattern}{A pattern to filter the results.}

\item{theme}{Name of the theme for which to search the variables.}
}
\value{
a \code{data.frame} with two variables:
\itemize{
 \item{\code{variable}: }{name of the variable.}
 \item{\code{value}: }{default value used in theme.}
}
}
\description{
Search Bootstrap variables
}
\examples{

# List default variables for Bootstrap 3
search_vars_bs()

# Variables for flatly theme
search_vars_bs("flatly")

}
