% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-vars.R
\name{search_vars_bs4dash}
\alias{search_vars_bs4dash}
\title{Search bs4Dash variables}
\usage{
search_vars_bs4dash(pattern = NULL, source = c("adminlte", "bootstrap"))
}
\arguments{
\item{pattern}{A pattern to filter the results.}

\item{source}{Search variables in AdminLTE or Bootstrap or both.}
}
\value{
a \code{data.frame} with three columns:
\itemize{
\item \code{source} : AdminLTE or Bootstrap variable.
\item \code{variable} : name of the variable.
\item \code{value} default value used.
}
}
\description{
Search bs4Dash variables
}
\examples{
# Retrieve all variables
all_vars <- search_vars_bs4dash()
head(all_vars, 20)

# Search for a pattern
head(search_vars_bs4dash("navbar"))
}
