% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pretty.R
\name{use_pretty}
\alias{use_pretty}
\title{Use pretty-checkbox CSS in Shiny}
\usage{
use_pretty(path)
}
\arguments{
\item{path}{Path to the file created with \code{\link{create_pretty}},
the file must be in www/ directory of the application.}
}
\description{
After created new pretty-checkbox CSS with \code{\link{create_pretty}},
 allow to use in Shiny application instead of default shinyWidgets dependency.
}
