% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars-scss.R
\name{vars_nav}
\alias{vars_nav}
\title{CSS navbar variables}
\usage{
vars_nav(height = NULL, default_color = NULL, default_bg = NULL,
  default_border = NULL, default_link_color = NULL,
  inverse_color = NULL, inverse_bg = NULL, inverse_border = NULL,
  inverse_link_color = NULL)
}
\arguments{
\item{height}{Height of the navbar, e.g. \code{"50px"} (the default in Bootstrap).}

\item{default_color}{Color of text in the navbar.}

\item{default_bg}{Background color of the navbar.}

\item{default_border}{Border color of the navbar.}

\item{default_link_color}{Link color.}

\item{inverse_color}{Color of text for inversed navbar.}

\item{inverse_border}{Border color for inversed navbar.}

\item{inverse_link_color}{Link color for inversed navbar.}

\item{inverse_bgBackground}{Background color for inversed navbar.}
}
\value{
a \code{list} that can be used in \code{\link{create_theme}}.
}
\description{
Those variables can be used to customize Bootstrap and Bootswatch themes.
}
