% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars-adminlte.R
\name{adminlte_global}
\alias{adminlte_global}
\title{AdminLTE CSS global variables}
\usage{
adminlte_global(content_bg = NULL, box_bg = NULL, info_box_bg = NULL)
}
\arguments{
\item{content_bg}{Background color of the body.}

\item{box_bg}{Default background color for boxes.}

\item{info_box_bg}{Default background color for info boxes.}
}
\value{
a \code{list} that can be used in \code{\link{create_theme}}.
}
\description{
Those variables can be used to customize
 global settings in {shinydashboard}.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinydashboard)

  ui <- dashboardPage(
    header = dashboardHeader(title = "My dashboard"),
    sidebar = dashboardSidebar(),
    body = dashboardBody(

      use_theme(create_theme(
        adminlte_global(
          content_bg = "#FAAC58"
        )
      ))
    )
  )

  server <- function(input, output, session) {

  }

  shinyApp(ui, server)
}
}
