% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars-bs4dash.R
\name{bs4dash_status}
\alias{bs4dash_status}
\title{bs4Dash status colors}
\usage{
bs4dash_status(
  primary = NULL,
  secondary = NULL,
  success = NULL,
  info = NULL,
  warning = NULL,
  danger = NULL,
  light = NULL,
  dark = NULL
)
}
\arguments{
\item{primary}{Default: \Sexpr[results=rd, stage=render]{fresh:::rd_color_tag("#0073b7")}}

\item{secondary}{Default: \Sexpr[results=rd, stage=render]{fresh:::rd_color_tag("#6c757d")}}

\item{success}{Default: \Sexpr[results=rd, stage=render]{fresh:::rd_color_tag("#28a745")}}

\item{info}{Default: \Sexpr[results=rd, stage=render]{fresh:::rd_color_tag("#17a2b8")}}

\item{warning}{Default: \Sexpr[results=rd, stage=render]{fresh:::rd_color_tag("#ffc107")}}

\item{danger}{Default: \Sexpr[results=rd, stage=render]{fresh:::rd_color_tag("#dc3545")}}

\item{light}{Default: \Sexpr[results=rd, stage=render]{fresh:::rd_color_tag("#f8f9fa")}}

\item{dark}{Default: \Sexpr[results=rd, stage=render]{fresh:::rd_color_tag("#343a40")}}
}
\value{
a \code{list} that can be used in \code{\link{create_theme}}.
}
\description{
bs4Dash status colors
}
\examples{
examples/bs4dash_status.R
}
