% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars-bs4dash.R
\name{bs4dash_yiq}
\alias{bs4dash_yiq}
\title{bs4Dash color contrast}
\usage{
bs4dash_yiq(contrasted_threshold = NULL, text_dark = NULL, text_light = NULL)
}
\arguments{
\item{contrasted_threshold}{The yiq lightness value that
determines when the lightness of color changes from "dark"
to "light". Acceptable values are between 0 and 255.}

\item{text_dark}{Dark text color.}

\item{text_light}{Light text color.}
}
\value{
a \code{list} that can be used in \code{\link{create_theme}}.
}
\description{
These variables allow to customize color used if
contrast between a color and its background is under threshold.
For example, it's used to choose text color written in \code{bs4ValueBox}
with background defined by a status.
}
