% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars-scss.R
\name{bs_vars_global}
\alias{bs_vars_global}
\title{Bootstrap CSS global variables}
\usage{
bs_vars_global(body_bg, text_color, link_color, link_hover_color,
  line_height_base, grid_columns, grid_gutter_width)
}
\arguments{
\item{body_bg}{Background color for the body.}

\item{text_color}{Global text color on body.}

\item{link_color}{Global textual link color.}

\item{link_hover_color}{Link hover color.}

\item{line_height_base}{Unit-less `line-height` for use in components like buttons.}

\item{grid_columns}{Number of columns in the grid, e.g. in \code{\link[shiny]{column}}.}

\item{grid_gutter_width}{Padding between columns. Gets divided in half for the left and right.}
}
\value{
a \code{list} that can be used in \code{\link{create_theme}}.
}
\description{
Those variables can be used to customize
 Bootstrap and Bootswatch themes.
}
\examples{

bs_vars_global(
  body_bg = "#FAFAFA" # change background color
)
}
