% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_theme.R
\name{create_theme}
\alias{create_theme}
\title{Create a custom Bootstrap theme}
\usage{
create_theme(theme = c("default", "cosmo", "flatly"), ...,
  output_file = NULL)
}
\arguments{
\item{theme}{Base theme to use.}

\item{...}{Lists of CSS variables declared with \code{\link{bs_vars_color}},  \code{\link{bs_vars_nav}}.}

\item{output_file}{Specifies path to output file for compiled CSS.}
}
\description{
Allow to customize some CSS variables from Bootstrap themes to be included in Shiny applications.
}
