% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_theme.R
\name{create_theme}
\alias{create_theme}
\title{Create a custom Bootstrap theme}
\usage{
create_theme(..., theme = c("default", "cerulean", "cosmo", "cyborg",
  "darkly", "flatly", "journal", "lumen", "paper", "readable", "sandstone",
  "simplex", "slate", "spacelab", "superhero", "united", "yeti"),
  output_file = NULL)
}
\arguments{
\item{...}{Lists of CSS variables declared with \code{bs_vars_*} or \code{adminlte_*} functions.}

\item{theme}{Base theme to use.}

\item{output_file}{Specifies path to output file for compiled CSS.}
}
\value{
If \code{output_file = NULL}, the function returns a string value of the compiled CSS.
 If the output path is specified, the compiled CSS is written to that file and \code{invisible()} is returned.
}
\description{
Allow to customize some CSS variables from Bootstrap themes to be included in Shiny applications.
}
\examples{

# Temporary file
tmp <- file.path(tempdir(), "my-pretty.css")

# Create the new theme
create_theme(
  theme = "default",
  bs_vars_color(
    brand_primary = "#75b8d1",
    brand_success = "#c9d175",
    brand_info = "#758bd1",
    brand_warning = "#d1ab75",
    brand_danger = "#d175b8"
  ),
  bs_vars_nav(
    default_bg = "#75b8d1",
    default_color = "#FFFFFF",
    default_link_color = "#FFFFFF",
    default_link_active_color = "#FFFFFF"
  ),
  output_file = tmp
)

# clean
unlink(tmp)
}
