% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_db_conn.R
\name{frs_db_conn}
\alias{frs_db_conn}
\title{Connect to FWA PostgreSQL Database}
\usage{
frs_db_conn(
  dbname = Sys.getenv("PG_DB_SHARE"),
  host = Sys.getenv("PG_HOST_SHARE"),
  port = Sys.getenv("PG_PORT_SHARE"),
  user = Sys.getenv("PG_USER_SHARE"),
  password = Sys.getenv("PG_PASS_SHARE")
)
}
\arguments{
\item{dbname}{Database name. Default: \code{Sys.getenv("PG_DB_SHARE")}.}

\item{host}{Host name. Default: \code{Sys.getenv("PG_HOST_SHARE")}.}

\item{port}{Port number. Default: \code{Sys.getenv("PG_PORT_SHARE")}.}

\item{user}{User name. Default: \code{Sys.getenv("PG_USER_SHARE")}.}

\item{password}{Password. Default: \code{Sys.getenv("PG_PASS_SHARE")}.}
}
\value{
A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object.
}
\description{
Opens a connection to a PostgreSQL database containing fwapg, bcfishpass,
and bcfishobs. Connection parameters default to environment variables
matching the \verb{PG_*_SHARE} convention used by fpr.
}
\examples{
\dontrun{
conn <- frs_db_conn()
DBI::dbDisconnect(conn)
}
}
\seealso{
Other database: 
\code{\link{frs_db_query}()}
}
\concept{database}
