% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_db_query.R
\name{frs_db_query}
\alias{frs_db_query}
\title{Query FWA PostgreSQL Database}
\usage{
frs_db_query(query, ...)
}
\arguments{
\item{query}{Character. SQL query string.}

\item{...}{Additional arguments passed to \code{\link[=frs_db_conn]{frs_db_conn()}}.}
}
\value{
An \code{sf} data frame (if the query returns geometry) or a plain
data frame.
}
\description{
Connects via \code{\link[=frs_db_conn]{frs_db_conn()}}, executes a SQL query, disconnects, and
returns the result. Uses \code{\link[sf:st_read]{sf::st_read()}} so spatial columns are returned
as sf geometry.
}
\examples{
\dontrun{
frs_db_query("SELECT * FROM whse_basemapping.fwa_lakes_poly LIMIT 5")
}
}
\seealso{
Other database: 
\code{\link{frs_db_conn}()}
}
\concept{database}
