% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_network_upstream.R
\name{frs_network_upstream}
\alias{frs_network_upstream}
\title{Get Stream Segments Upstream of a Network Position}
\usage{
frs_network_upstream(blue_line_key, downstream_route_measure, ...)
}
\arguments{
\item{blue_line_key}{Integer. Blue line key of the reference point.}

\item{downstream_route_measure}{Numeric. Downstream route measure of the
reference point.}

\item{...}{Additional arguments passed to \code{\link[=frs_db_conn]{frs_db_conn()}}.}
}
\value{
An \code{sf} data frame of upstream stream segments.
}
\description{
Returns all FWA stream segments upstream of a given blue line key and
downstream route measure. Uses the fwapg \code{fwa_upstream()} ltree comparison.
}
\examples{
\dontrun{
# Get all streams upstream of a point on the Bulkley
upstream <- frs_network_upstream(
  blue_line_key = 360873822,
  downstream_route_measure = 166030
)
}
}
\seealso{
Other traverse: 
\code{\link{frs_network_downstream}()}
}
\concept{traverse}
