% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_point_locate.R
\name{frs_point_locate}
\alias{frs_point_locate}
\title{Locate a Point on the FWA Stream Network}
\usage{
frs_point_locate(conn, blue_line_key, downstream_route_measure)
}
\arguments{
\item{conn}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object (from \code{\link[=frs_db_conn]{frs_db_conn()}}).}

\item{blue_line_key}{Integer. Blue line key of the stream.}

\item{downstream_route_measure}{Numeric. Downstream route measure in metres.}
}
\value{
An \code{sf} data frame with a single point geometry.
}
\description{
Given a blue line key and downstream route measure, return the point
geometry on the stream network. Wraps fwapg \code{fwa_locatealong()}.
}
\examples{
\dontrun{
conn <- frs_db_conn()
pt <- frs_point_locate(conn, blue_line_key = 360873822,
  downstream_route_measure = 1000)
DBI::dbDisconnect(conn)
}
}
\seealso{
Other index: 
\code{\link{frs_point_snap}()}
}
\concept{index}
